Add frictional loss calibration and correction
This code still needs significant cleaning. Now individual test methods can be called by a string at the start of the code (e.g., VCC_CONSTANT_RATE). A bug was also corrected in the overload protection thread that resulted in an overflowing buffer when the rate was high enough to fill the buffer in the test. Finally, many small plotting functions and other misc. quality of life improvements were added.